Ungerade Funktion implementiert
- in excel.ts ungeradeFragment hinzugefügt
-  in excel-constants ungerade funktion herausgenommen
-ungerade funktion in ungerade.ts ausgegliedert
